pyproject.toml
[project]
name = "psiutils"
dynamic = ["version"]
description = "Various TKinter utilities."
authors = [{name = "Jeff", "email" = "<jeffwatkins2000@gmail.com>"}]
requires-python = '>= 3.10, < 3.13'
readme = "README.md"

include = ["HISTORY.md", "src/psiutils/icons/*.png"]

dependencies = [
    "appdirs>=1.4.4",
    "iniconfig>=2.1.0",
    "markdown>=3.8.2",
    "packaging>=25.0",
    "pillow>=11.3.0",
    "pluggy>=1.6.0",
    "psi-toml>=0.0.12",
    "psiconfig>=0.0.14",
    "pycairo>=1.28.0",
    "pygments>=2.19.2",
    "pygobject>=3.54.2",
    "pytest>=9.0.2",
    "structlog>=25.4.0",
    "tkcalendar>=1.6.1",
    "tkinterweb>=4.4.4",
    "tkinterweb-tkhtml>=1.1.1",
    "tomli>=2.2.1",
    "tomli-w>=1.2.0",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/psiutils/_version.py"

[dependency-groups]
dev = ['pytest']

[build-system]
requires = ['uv-build']
build-backend = 'uv_build'
